# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WFoods ComandaFácil API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginBody {
  email: string;
  password: string;
  tenantSlug: string;
}

export interface WaiterLoginBody {
  name: string;
  password: string;
  tenantSlug: string;
}

export type CurrentUserRole =
  (typeof CurrentUserRole)[keyof typeof CurrentUserRole];

export const CurrentUserRole = {
  admin: "admin",
  waiter: "waiter",
} as const;

export interface CurrentUser {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  role: CurrentUserRole;
  tenantId: number;
  tenantSlug: string;
  tenantName: string;
}

export interface AuthResponse {
  token: string;
  user: CurrentUser;
}

export interface Tenant {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  active: boolean;
  createdAt: string;
}

export interface CreateTenantBody {
  name: string;
  slug: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  adminEmail: string;
  adminPassword: string;
}

export interface UpdateTenantBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  active?: boolean | null;
}

export interface Category {
  id: number;
  tenantId: number;
  name: string;
  sortOrder: number;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  sortOrder?: number;
}

export interface UpdateCategoryBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  sortOrder?: number | null;
}

export interface Product {
  id: number;
  tenantId: number;
  /** @nullable */
  categoryId?: number | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  available: boolean;
  /** @nullable */
  printSector?: string | null;
  createdAt: string;
}

export interface CreateProductBody {
  /** @nullable */
  categoryId?: number | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  available?: boolean;
  /** @nullable */
  printSector?: string | null;
}

export interface UpdateProductBody {
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  available?: boolean | null;
  /** @nullable */
  printSector?: string | null;
}

export interface PublicMenu {
  tenant: Tenant;
  categories: Category[];
  products: Product[];
}

export interface Table {
  id: number;
  tenantId: number;
  number: number;
  /** @nullable */
  label?: string | null;
  active: boolean;
  createdAt: string;
}

export interface CreateTableBody {
  number: number;
  /** @nullable */
  label?: string | null;
}

export interface UpdateTableBody {
  /** @nullable */
  number?: number | null;
  /** @nullable */
  label?: string | null;
  /** @nullable */
  active?: boolean | null;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  printSector?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  novo: "novo",
  preparando: "preparando",
  pronto: "pronto",
  finalizado: "finalizado",
} as const;

export interface Order {
  id: number;
  tenantId: number;
  /** @nullable */
  tableId?: number | null;
  /** @nullable */
  tableNumber?: number | null;
  /** @nullable */
  waiterId?: number | null;
  /** @nullable */
  waiterName?: string | null;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderItemBody {
  productId: number;
  quantity: number;
  /** @nullable */
  notes?: string | null;
}

export interface CreateOrderBody {
  tenantSlug: string;
  /** @nullable */
  tableId?: number | null;
  /** @nullable */
  customerName?: string | null;
  items: CreateOrderItemBody[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  novo: "novo",
  preparando: "preparando",
  pronto: "pronto",
  finalizado: "finalizado",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

export const PaymentMethod = {
  dinheiro: "dinheiro",
  pix: "pix",
  cartao: "cartao",
} as const;

export interface Payment {
  id: number;
  tenantId: number;
  orderId: number;
  amount: number;
  method: PaymentMethod;
  createdAt: string;
}

export type CreatePaymentBodyMethod =
  (typeof CreatePaymentBodyMethod)[keyof typeof CreatePaymentBodyMethod];

export const CreatePaymentBodyMethod = {
  dinheiro: "dinheiro",
  pix: "pix",
  cartao: "cartao",
} as const;

export interface CreatePaymentBody {
  orderId: number;
  amount: number;
  method: CreatePaymentBodyMethod;
}

export type DashboardSummaryRevenueByMethod = {
  dinheiro: number;
  pix: number;
  cartao: number;
};

export interface OrdersByStatus {
  novo: number;
  preparando: number;
  pronto: number;
  finalizado: number;
}

export interface DashboardSummary {
  totalOrdersToday: number;
  totalRevenueToday: number;
  activeOrders: number;
  ordersByStatus: OrdersByStatus;
  revenueByMethod: DashboardSummaryRevenueByMethod;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalSold: number;
  totalRevenue: number;
}

export interface Waiter {
  id: number;
  tenantId: number;
  name: string;
  createdAt: string;
}

export interface CreateWaiterBody {
  name: string;
  password: string;
}

export type ListProductsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  available?: boolean | null;
};

export type ListOrdersParams = {
  /**
   * @nullable
   */
  status?: ListOrdersStatus;
  /**
   * @nullable
   */
  tableId?: number | null;
  /**
   * @nullable
   */
  date?: string | null;
};

export type ListOrdersStatus =
  | (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus]
  | null;

export const ListOrdersStatus = {
  novo: "novo",
  preparando: "preparando",
  pronto: "pronto",
  finalizado: "finalizado",
} as const;

export type ListPaymentsParams = {
  /**
   * @nullable
   */
  date?: string | null;
};

export type GetTopProductsParams = {
  /**
   * @nullable
   */
  limit?: number | null;
};