# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WFoods ComandaFácil API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/waiter/login:
    post:
      operationId: waiterLogin
      tags: [auth]
      summary: Waiter login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaiterLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthorized
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories for current tenant
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{categoryId}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Products
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current tenant
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: ["novo", "preparando", "pronto", "finalizado", null]
        - name: tableId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (client or waiter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Payments
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments for current tenant
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Register a payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # Dashboard / Analytics
...
components:
  schemas:
...
    AdminLoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        tenantSlug:
          type: string
      required:
        - email
        - password
        - tenantSlug
    WaiterLoginBody:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
        tenantSlug:
          type: string
      required:
        - name
        - password
        - tenantSlug
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/CurrentUser"
      required:
        - token
        - user
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, waiter]
        tenantId:
          type: integer
        tenantSlug:
          type: string
        tenantName:
          type: string
      required:
        - id
        - name
        - role
        - tenantId
        - tenantSlug
        - tenantName
...
    Category:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - name
        - sortOrder
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
      required:
        - name
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        sortOrder:
          type: ["integer", "null"]
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        notes:
          type: ["string", "null"]
        printSector:
          type: ["string", "null"]
      required:
        - id
        - orderId
        - productId
        - productName
        - quantity
        - unitPrice
    Order:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        tableId:
          type: ["integer", "null"]
        tableNumber:
          type: ["integer", "null"]
        waiterId:
          type: ["integer", "null"]
        waiterName:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [novo, preparando, pronto, finalizado]
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenantId
        - status
        - total
        - items
        - createdAt
        - updatedAt
    CreateOrderBody:
      type: object
      properties:
        tenantSlug:
          type: string
        tableId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
      required:
        - tenantSlug
        - items
    CreateOrderItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        notes:
          type: ["string", "null"]
      required:
        - productId
        - quantity
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [novo, preparando, pronto, finalizado]
      required:
        - status
    Payment:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        orderId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [dinheiro, pix, cartao]
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - orderId
        - amount
        - method
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        orderId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [dinheiro, pix, cartao]
      required:
        - orderId
        - amount
        - method
...