# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WFoods ComandaFácil API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  tenantId: number;
  /** @nullable */
  tableId?: number | null;
  /** @nullable */
  tableNumber?: number | null;
  /** @nullable */
  waiterId?: number | null;
  /** @nullable */
  waiterName?: string | null;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  createdAt: string;
  updatedAt: string;
}